Fix bit test for if a frame is initialized
Summary:
When an inlined function unwinds we can either just unlink it from the thread state or if the frame has been reified we need to copy the frame state into the reified frame. This latter part happens via `JITRT_UnlinkFrame`. We're not currently doing the right thing to successfully call `JITRT_UnlinkFrame`. We test the wrong bit, and we expect it to be the wrong value.

This tests the right bit and does the right condition.

Reviewed By: alexmalyshev

Differential Revision: D80828825

fbshipit-source-id: e8a86802e608c6eb9fdf7fbe763238c45faf80ae