TO_BOOL optimizations
Summary:
There's a number of optimizations that can be applied to `TO_BOOL`.

If we're doing a `LOAD_CONST` and then `TO_BOOL` it can be calculated statically.

If we're doing a compare then there's a special form that ensures a bool is pushed instead of doing the conversion in the byte code.

Reviewed By: alexmalyshev

Differential Revision: D80763700

fbshipit-source-id: 4ad34ff26055b4ddb1ec7c3c9184d3bb2d9112dc