Fix JUMP optimizations
Summary: There's new pseudo-ops for jumps which are `JUMP_IF_TRUE`/`JUMP_IF_FALSE` which will expand to a `TO_BOOL` + `JUMP_IF` form when left in. We need to thread jumps through these and we also need to do the load const is jump optimization around a `TO_BOOL`.

Reviewed By: jbower-fb

Differential Revision: D80820423

fbshipit-source-id: 5abc2bb3db4bf802989be172ab3c2ccb8343f37e