Flowing jumps replaces with NOP and then adds new instruction
Summary:
https://github.com/python/cpython/commit/0315941441f1e5f944a758c67eb1763b00e6e462

This is a subtle change in the code generation. Previously the existing instruction was updated, but now the instruction is NOP'd and a new one is appened.

This NOP and then replace results in losing the exception handling information for the new jump and throws off the exception handling tables. We match the behavior.

Reviewed By: martindemello

Differential Revision: D80987789

fbshipit-source-id: c7fb0f7799ac58ea6bf2d134239da8b536fefc5e