Simplify how lir::Instruction manages its inputs
Summary:
There's a number of places where inputs could be set directly, but are currently
being removed and appended instead.  Make it easier to construct operands by not
requiring them to know their "owning" instruction upfront, which then makes it
easier to set them on instructions.

Remove vestiges of def-use chains on instructions, all we need is
LinkedOperand::setLinkedInstr().

Reviewed By: martindemello

Differential Revision: D81802873

fbshipit-source-id: c955fd91cc2bf707f59af70c85de688d21e60605