Shrink the size of PhyLocation
Summary:
It's currently an `int` and a `size_t`.  With padding that sums to 16 bytes.
There's no need to use a `size_t` to represent the bit size of the location, it
maxes out at 128.  Use a `uint32_t` which drops the struct's size down to 8
bytes.  This also drops the size of Operand from 48 to 40 bytes.

Reviewed By: martindemello

Differential Revision: D81930040

fbshipit-source-id: 3facb4560b4ae02960707db8c4abb0faaeb51447