Don't propagate tstate past return
Summary:
Currently we're blowing up trying to keep something in rax alive past a return - that thing is the thread state, as it's the only thing that can be used after we "return". This makes it so we keep nothing alive past the return.

I think the core issue here is that the register allocator just has no clue what's going on... The return instruction takes an input which is the return value, but the register allocator doesn't know that we need to keep the input alive *past* the return. Probably the special treatment of return that we do isn't great either, but as long as nothing lives past return it's probably fine.

So this gets rid of the keeping things alive past return. Instead we'll grab the thread state when unlinking the frame. For 3.10 shadow frames this means always calling unlink frame and pushing the check to see if we have a reified frame into the unlinker.

Another approach would probably be to make return take 2 inputs - the returned value and the thread state. But it seems extra wonky. I think we should probably move towards having return be a real return, and then we can use LIR to generate all of the appropriate unwind logic.

Reviewed By: alexmalyshev

Differential Revision: D81733991

fbshipit-source-id: 2aec9599413cde591feef0c808f692da566afeef