Don't start new block for try body
Summary: We need to avoid making a new basic block to match CPython's behavior around load borrow optimizations which only work across the basic block. Ideally we should add a pass which re-constructs the basic blocks based upon the actual jump instructions (which is what CPython is doing) but we're generally not too far off from CPython's behavior elsewhere so this is an easy fix.

Reviewed By: alexmalyshev

Differential Revision: D81642076

fbshipit-source-id: 9311745405209c7d0700dadaa50e4862106d5a51