More borrow analysis fixes
Summary:
There's one thing the AI got wrong - the check to see if we're doing a load super attr isn't quite right

There's also an issue in the upstream analysis - the adding of refs for instructions which only produce things is adding a ref associated with the incorrect instruction. It shouldn't be the iteration for the number of pushed values, it should be for the current instruction. This will be fixed upstream with https://github.com/python/cpython/pull/138678 and backported to 3.14 with D81950973.

Reviewed By: mpage

Differential Revision: D81642077

fbshipit-source-id: 4c8fe7f33fe38fb23c1cbf83074649f272898f71