Create new block for final instructions inserted into a lambda
Summary:
CPython will usually insert a `LOAD_CONST`/`RETURN_VALUE` at the end of a lambda, and then this will get optimized away. This has a side effect of leaving a `None` const on the lambda.

We emit it in the same block currently which can prevent some other optimizations because we'll see the return block as being too large.

Reviewed By: jbower-fb

Differential Revision: D81977406

fbshipit-source-id: e6cccd87799330b3d65534f9ed6ed5c4c13b6b23