Support dynamically enabling the JIT with a new call count threshold
Summary:
Adds new Python API functions for configuring the JIT to compile functions
automatically.

`cinderx.jit.compile_after_n_calls(N)`: As it sounds, this tells the JIT to
compile functions after they have been called `N` times.  This is equivalent to
`PYTHONJITAUTO=N`, but it can be called multiple times as a program runs.

`cinderx.jit.auto()`: This tells the JIT to automatically compile functions, but
leaves all the decision-making to the JIT.  It's the recommended behavior for
users.

Reviewed By: jbower-fb

Differential Revision: D81356873

fbshipit-source-id: 94145aa1b4d00d29301ae8c5d9651c932e2e9584