Compute cold basic blocks
Summary: Previously we were able to skip calculation of cold basic blocks as they were identical to not warm blocks. But with the preservation of non-empty blocks in 3.14 this has changed and blocks can end up both not warm and not cold. We therefore need to calculate cold blocks now too.

Reviewed By: mpage

Differential Revision: D82127010

fbshipit-source-id: fc5f6c0f1981aaea6ad9472b4c236ecaae2127bd