Add JIT support for LOAD_BUILD_CLASS
Summary:
Apparently we've never supported this?! Quite surprising as it must mean we were at least missing quite a few functions from JIT during test.

The code assumes the builtins are always a dictionary, whereas the interpreters bytecode can handle a non-dictionary `PyObject*`. I think this assumption is safe as the preloader seems to always have the builtins as a dictionary.

Reviewed By: alexmalyshev

Differential Revision: D82689837

fbshipit-source-id: 6d10820eea803a365d14ec5268dbfe56aa6b502d