Disable JIT when sys.{monitoring,setprofile,settrace} are being used, re-enable afterwards
Summary:
Deopts all functions and disables the JIT when a sys.setprofile, sys.settrace,
or sys.monitoring callback is registered.  Re-enables the JIT once all the
callbacks are cleared out.

This patches the respective functions with ones that we provide.  This allows us
to catch when tracing/monitoring functionality is enabled/disabled and flip the
JIT to the matching state as well.

Ideally this would make use of the sys.audit API, but the events it exports
don't give you enough information to tell whether a callback is being registered
or removed.  That's a change we'd like to make in upstream.

Reviewed By: jbower-fb

Differential Revision: D83086287

fbshipit-source-id: 5d11418b3f4d57c0d8d910eee99445dcc61d9089