Reduce manually copied code from CPython for generators
Summary:
If we're a little careful we can avoid having to manually copy and tweak `gen_traverse()` and `gen_dealloc()`.

For `gen_dealloc()` we can actually just borrow and use the preprocessor to apply the critical change to considers our free-list on the final release of memory.

For `gen_traverse()` we mostly just need to traverse the references only the JIT will know about before deferring to the base functionality. One caveat here though is we need to have the `_PyInterpreterFrame` a bit more populated than lightweight frames do by default, otherwise the base traversing code may fail. We can gate populating these extra fields to be for generators only.

Reviewed By: DinoV

Differential Revision: D82487092

fbshipit-source-id: a36d9927edf10e7f8c5ee13b99e2982666578bc7