Improve how Python tests deal with Python versions
Summary:
Trying to keep each file to be owned by one rule, instead of having multiple
rules for different versions.  We already have the `python_test` macro for
multiplexing across Python versions, so make use of that.  It will also set
target platforms soon, so best to make use of it wherever we can.

Adding an `exclude_py_versions` attribute to the macro so that we'll
automatically pick up new Python versions as they're added to the macro, instead
of accidentally leaving some tests as `["3.10", "3.12"]` and not picking up 3.14
when it's eventually added.

Had to tweak the perf map test file to not use JitAll for its subprocesses.  It
seems to start flaking out more and compile way more than it expects, but it's
inconsistent across runs.  The test already uses `cinderx.jit.lazy_compile` to
select functions to compile, it doesn't need `-X jit-all`.

Reviewed By: jbower-fb

Differential Revision: D83368971

fbshipit-source-id: 7f0792dcbdd0c528ca741867bcc5b96aa92c93c5