Pre-populate jump ioparg with target instruction index
Summary:
CPython is doing a weird thing where it's using `i_oparg` to transfer the target instruction sequence to be resolved later, it then populates `i_target` with this info and doesn't clear `i_oaprg`.

This results in `i_oparg` having a number which will impact the resolution of the offsets in `resolve_jump_offsets`. It takes some very specific code but we can end up with a `EXTENDED_ARG`/`JUMP` pair where the oparg is 256.

If the EXTENDED_ARG were to be removed this would fit in just the JUMP oparg, but there's no logic to do this specialized shrink.

Reviewed By: alexmalyshev

Differential Revision: D82142450

fbshipit-source-id: dc67a361b968b5c547824407aa7b1bd1db38e6db