Disable Imports Monitor on Cinder jit test
Summary:
Ensure that the Cinder jit test `test_max_code_size_slow` passes with Imports Monitor enabled (as we plan on Imports Monitor being enabled for all fbcode linux tests). This test broke on D82670645.

The test tests the max code size, which seems to be getting thrown off with the inclusion of Imports Monitor code. Passing in the `-S` flag to the `subprocess.run` command disables the import of the `site` module. This module imports `sitecustomize` which is where startup functions like Imports Monitor and Lazy Imports are run. So by using `-S` we avoid adding Imports Monitor to the test.

Note that this doesn't fix the root issue of the test being unstable, but a task (T240152676) has been filed for this.

Reviewed By: alexmalyshev

Differential Revision: D83766379

fbshipit-source-id: c6bafecbcaac32896ccba861493f7a4eb84e30db