Return UNKNOWN_ERROR when compile_func() can't find any preloaders
Summary:
This is a rare case, being triggered in
https://fb.workplace.com/groups/fbpython/posts/31794532546828042.  A function is
being compiled (`libfb.py.sitevar:get_sitevar`), it goes through preloading and
turns into a batch of 2 functions (`[get_sitevar, get_sitevars]`) but then both
preloaders cannot be found, presumably because they were deleted as a side
effect of preloading.  This ends up returning an uninitialized _PyJIT_Result
value and we later crash trying to verify it.

Always default the return code for this kind of case to UNKNOWN_ERROR.

Reviewed By: itamaro

Differential Revision: D83843993

fbshipit-source-id: 1547bd59a8253fc1bbb09444b4c4cb0e87b80963