Abstract DeoptPatcher into CodePatcher
Summary:
Making a more general abstraction that can patch arbitrary sections of code.  Or
at least up to 7 bytes :).

The CodePatcher stores a small buffer of data to patch in plus the destination
address.  When a patch is triggered, it will swap what it has in its buffer with
what is at the destination address.  This allows the CodePatcher to undo its
patch in the future with an "unpatch".

A "deopt patcher" is a specific case of this, and there's now a new subclass
called JumpPatcher that covers this case, but it's still more general in that
it's focused on patching in jumps rather than anything specific to deopts.

The 7-byte limit here is to keep the CodePatcher to a small size, currently 24
bytes.  If in the future we'll want to patch in 8-byte values we can replace its
vtable with a smaller enum value and shrink it further.

Reviewed By: jbower-fb

Differential Revision: D83591551

fbshipit-source-id: 160023a047397fb9ef21452e6914bcc49a1ee95d