Fix opt interpreter build w/ static opcodes being present.
Summary: The opt build of the interpreter is a little messed up when we actually try and use static Python stuff. That's because the generated cases get included earlier (when we include ceval.c). This moves the include down to after the static Python definitions and gets rid of the ones which use `GETITEM` which won't be defined (but we don't need anymore on 3.14).

Reviewed By: jbower-fb

Differential Revision: D84014691

fbshipit-source-id: d6dcea4f6374b74435f5998b733963f4f09757a6