Assign extended opcode numbers
Summary:
Updates the assignment script to assign new IDs to the static opcodes on 3.14. These all get assigned w/ an opcode ID that tries to be similar to a base opcode id. That lets the disassembler do an okay-ish job when displaying them in that we'll see `EXTENDED_OPCODE` and then the argument is properly displayed afterwards.

In our opcode definitions we also need to avoid trampling on the normal opcode state for several cases. So we now have our own `STATIC_OPMAP`, `STATIC_OPNAMES` and `STATIC_CONST_OPCODES` to store the specific information about static Python. We also need to track the consts by names to avoid looking at the same opcode # and getting the wrong comparison.

Reviewed By: jbower-fb

Differential Revision: D83679258

fbshipit-source-id: bba4bca6f5e26e32d5eaa920802ab020d982345d