Fix test_cinderx.test_immortalize and test_cinderx.test_jit_global_cache
Summary:
These tests are failing because multiprocessing.Process has switched from using fork to forkserver by default. This primarily fixes these tests by switching the default.

When the current tests fail they actually hang until the timeout though... this fixes that so they raise an exception.

This also exposed that our current tests in test_cinderx.test_jit_global_cache which depend on skipping due to lazy imports are actually never running as the decorator is applied incorrectly. This fixes the usage of the decorator and some things that have broken in the tests over time.

Reviewed By: alexmalyshev

Differential Revision: D84379796

fbshipit-source-id: 6d43fa546c1d6e0c48c4eeac138acbfc7551ca74