Elimiante strict module's builtins isolation
Summary:
Strict modules originally supported builtins isolation so that they wouldn't see any changes to builtins. This guaranteed that the analysis always knew what the builtins where and that analysis wouldn't be invalidated by changes to builtins.

This is no longer relevant now that we've dropped the analyzer, and it comes with a cost of needing to check if a global value is actually defined in the module.

Reviewed By: alexmalyshev

Differential Revision: D84304584

fbshipit-source-id: 72e4eaf7d981740b75d0073f2cf9326f13f9fca8