Fix inline caches for module attributes on 3.14
Summary: In 3.14 we've lost the dictionary version number and now only have a keys version. The keys version only gets bumped when the actual layout of the keys have changed - simply updating a value doesn't cause it to update. This switches the module level caches on 3.14 to use the global cache / dictionary watchers instead.

Reviewed By: jbower-fb

Differential Revision: D84308698

fbshipit-source-id: 8052667177b5957e76102da2f236a372764e210b