Close parent perf map before copying
Summary:
In 3.12 the parent perf map would be closed by CPython after fork. In 3.14 that doesn't happen and we go and copy our perf map entries into it which never completes because we keep copying what we've copied.

This changes it so we explicitly close it and then it'll get re-opened.

https://github.com/python/cpython/commit/a73aa48e6bec900be7edd3431deaa5fc1d809e6f

Reviewed By: alexmalyshev

Differential Revision: D84396514

fbshipit-source-id: e9c9882ff96ccf700ae2af03635e453b8723f266