Move DeoptMetadata to CodeRuntime
Summary:
DeoptMetadata objects are tied to specific compiled code objects, but are
currently stored as a big vector in the Runtime singleton.  Without doing a lot
of memory shifting, they cannot be deleted when their corresponding compiled
code object is deleted so they effectively live forever.  And in the future,
accessing them will require taking a global lock when we support free-threading.

The simpler solution is to store them alongside other code object data in the
CodeRuntime.  This ties their lifetime closer to the rest of the compiled code
object.  It doesn't solve the deletion story _yet_ because we don't delete
CodeRuntime objects either (they're stored in a big arena), but we'll get there
eventually.  But more importantly it removes more usage of the global Runtime
singleton and avoids the need for future global locks.

The deopt stats objects are left alone in this diff.  We don't want to naively
move them into the CodeRuntime inline because they're mutable during code
execution, they would hit copy-on-write.

Reviewed By: DinoV

Differential Revision: D84292868

fbshipit-source-id: 5b6a8ed42c94dcfd22cfa551fdd3cf0229765b5e