Move deopt stats to CodeRuntime
Summary:
Like deopt metadata before them.  This is so compiled code objects own more of
their own memory and we can both deallocate it more easily in the future and
lock it more easily once we need to think about free-threading.

Care is taken to avoid copy-on-write when these are modified, they are stored
out-of-line.

Reviewed By: jbower-fb

Differential Revision: D84505937

fbshipit-source-id: 1b006b7752c6dbe3685c607271a68d70be4bc8cb