Update test_code_sbs for 3.14
Summary:
Lots of little changes to the code gen... This splits the tests which have changed into 3.10/3.12/3.14 versions and updates the 3.14 versions for those that are needed.

I'm tempted to just get rid of these tests as we didn't use them to drive the upgrade and instead used the test corpus to do so - and that has a lot more coverage.

Reviewed By: mpage

Differential Revision: D84185708

fbshipit-source-id: 0553fe3302e23298530daffd0289cd626d32c294