Mark x86-specific code regions
Summary:
Mark all of the places in cinder JIT where we are calling x86-specific APIs from asmjit. this introduces a new `arch.h` header, which provides a namespace that we now use, so that we can add additionally architectures without having to switch on macros so much.

Ideally, most of these can be abstracted, but there are some cases where checking macros will be more worth it in terms of maintainability, since they are very call-site specific.

Reviewed By: alexmalyshev

Differential Revision: D84636348

fbshipit-source-id: 881bf9149b8122930b5f2b24434cc0eec7818598