Fix ToBool/Make IsTruthy type CBool
Summary:
Fixes a bug using `IntConvert` on the output of `IsTruthy`. The `IntConvert` just cuts off the upper bits, which doesn't doesn't work when the length (32 bits) of a list is actually the value returned by `IsTruthy`.

To fix this I've gone and actually made `IsTruthy`'s output type `CBool` rather than `CInt32`. I don't see any reason we should be returning anything other than a bool. My suspicion is this was a mistake/historical in order to automatically get a deopting guard inserted for the case where the result is `-1` due to an error in one of the functions called by `IsTruthy`. We can just explicitly create the right guard in the internal implementation of `IsTruthy` instead.

As part of doing this I've also made `CompareBool` output type `CBool` as there are some cases where we simplify `IsTruthy` to `CompareBool` and we need to maintain the type.  It probably used `CInt32` for the same reason as `IsTruthy`.

It's not strictly needed, but I've also added an LIR instruction `NBitTo8BitBool`, which emits:
```
TEST r, r
SETNE r
```
In retrospect we could have used a LIR `Compare 0, r`, but I've implemented this now and it produces slightly shorter code.

Reviewed By: alexmalyshev

Differential Revision: D84438801

fbshipit-source-id: 8882e2e0d21cf610ea1d079c44f9c79a64ac5630