Don't re-compile to hit steady state in ref leak tests
Summary: If we jit every single time we run the test we won't hit the happy steady state as the JITed functions still keep things alive. We should at some point aspire to not have that be the case but for now let's just make sure we JIT each function once.

Reviewed By: alexmalyshev

Differential Revision: D84864693

fbshipit-source-id: ec63406c27661b7b402417ebaa45f668c88634d0