Remove sleep calls and extra macOS message in build
Summary:
The sleep calls are more annoying than they are useful right now.  The big
message about this being experimental is still kept, so we can still point
people to it.

The macOS message is kind of extra right now.  We build on macOS but as there's
no aarch64 backend for the JIT yet it's just a nothingburger of a Python
extension.

Reviewed By: DinoV

Differential Revision: D84923617

fbshipit-source-id: 032f075f59e526e50425453fff46c57f90786110