Add basic JIT tests to GitHub action
Summary:
Trying to use OSS tooling to actually run JIT tests.  The heavier tests like
`test_cinderjit.py` aren't compatible with OSS yet, mostly because of import
paths assuming we're in fbcode, but there's a few of the simpler ones here that
run out of the box.

This is definitely not the best way to orchestrate this, open to suggestions.
I'd like to get _something_ checked in, because right now we're only testing if
the module is able to initialize or not.

Reviewed By: yoney

Differential Revision: D85053738

fbshipit-source-id: 0ab549a9f63fd38541615e47365247cf843b828e