Fix memory leak in CAST opcode implementation
Summary: The CAST_COERCE_OR_ERROR is problematic in 3.14. We can't include a DECREF_INPUTS in there (it's not a real macro and gets expanded by the cases generator). We also need to be more careful about the int -> float case and not end up with 2 references to the new float.

Reviewed By: alexmalyshev

Differential Revision: D84867339

fbshipit-source-id: bfbcad6cc3ae7bba0f7ac4bb0ac6e4916cc05faa