Don't clear global caches in clear caches
Summary: This causes some new tests for 3.14 to fail but doesn't cause failures elsewhere. These caches are a little different in that they don't actually hold onto any references to any objects so not clearing them is fine. But clearing them causes functions to de-optimize in ref leak tests when the caches are empty.

Reviewed By: alexmalyshev

Differential Revision: D85256900

fbshipit-source-id: e8e8bbd267935a697aa779da769eb100d025e242