Fix static jump offsets
Summary:
The calculation for these is currently wrong in the compiler and the interpreter. The `EXTENDED_OPCODE` is already accounted for in the pc by the instr size, but the interpreter needs to jump over it.

The JIT is seeing the correct offset for the target and ends up parsing the byte code incorrectly.

Currently we have an extra `SKIP_OVER(1);` when the jump isn't taken which handles the cache size, and then a 2nd `SKIP_OVER(1);` for handling the `EXTENDED_OPCODE`. This moves to doing one `SKIP_OVER(2);` before the jump and then we either jump or dispatch.

Reviewed By: jbower-fb

Differential Revision: D85290581

fbshipit-source-id: 2041f7cb9d9d654e2d7a93cac312f03240fd1de5