Up/downgrade SplitMutator inline cache accessors as known values change
Summary: It's possible to get back to the containing `AttributeMutator` from a `SpltMutator` pointer as one contains the other, and there are also a few slots remaining in the 3-bit space of the `AttributeMutator::Kind`. Using this we can very cheaply add in some slightly more specialized behaviors when getting/setting instance attributes.

Reviewed By: DinoV

Differential Revision: D85970948

fbshipit-source-id: f893a503f814a9f4e2523c79d6d529fb293aa926