Match new line propagation
Summary:
I don't think we have any observable code differences here yet but it's possible this introduce some... In 3.15 the compiler will look for the next non-empty block when propagating line numbers, where as previously it would only look at the next block. This adds that change and makes the 3.15 version like upstream.

https://github.com/python/cpython/commit/e92599e677de8195898e686b61e930baa3b8f627

Reviewed By: martindemello

Differential Revision: D85973561

fbshipit-source-id: b99af2ec392276bceeda4a55f241b166dd1a41dc