Make all HIR .txt tests not skip and automatically update their expected output
Summary:
Apparently we just skip if there isn't expected output for a new version of Python.

To fix this I altered the test loader so it would output failed results (rather than skips) for tests with missing expected output for the current Python version. This way we can use the existing `update_hir_expected.py` script. As there are a small number of tests which we do actually want to skip in later Python versions I've added a new `--- Skip ---` directive which can be placed after any `--- Expected X.Y ---` sections to prevent the test being flagged as failed in future Python versions.

Reviewed By: DinoV

Differential Revision: D86166766

fbshipit-source-id: 3d912203c9de1e3cbf51b151197164da197506e9