Fix profiling skip
Summary:
We have a `test_profile` skip in `cinder_jit_ignore_tests.txt` - but I'm pretty sure it doesn't do anything as it's not structured as a regex (or maybe it only does something on 3.10).

The `test_profile` test is passing on 3.12 and 3.14 but new tests are failing on 3.15 so this disables them.

Reviewed By: yoney

Differential Revision: D86138385

fbshipit-source-id: 9a86f6d79bd735e4708f72ae91e1f46fd00a2c0c