Support aarch64 symbol identification in the disassembler
Summary:
When running the disassembler, x86-64 looks for call instruction operands that are absolute addresses and attempts to idnetify the function being called.

We want the same behavior in aarch64, but we don't have absolute addressing. Instead, when we find a br/blr instruction, walk backward through the instructions and attempt to build up the pointer by emulating the mov/movz/movk operations.

Reviewed By: mpage

Differential Revision: D86323380

fbshipit-source-id: bf9df0940e02781e4b0ff63fd2deb69ad6f23c53