Declare GDB JIT interface symbols as weak
Summary:
See https://fb.workplace.com/groups/pyupgrades/posts/3808916836079107.  CinderX
defines these symbols and so does LLVM ORC, which leads the linker to die with
duplicate symbol errors.

This fixes the build but it doesn't mean we're going to be able to use the
interface across both codebases.  It's not thread-safe.

Reviewed By: Yhg1s, itamaro

Differential Revision: D87075835

fbshipit-source-id: d06713fa473260232d9c89b09f3f8530f6a15c5c