Support querying/removing last register in sets
Summary:
When we're generating prologues and epilogues, we push and pop all of the necessary registers. When we pop, we pop in reverse order to make sure they get the correct values.

In order to accomplish this, we pull from the front in the prologue, and in the epilogue we pull from the front and create a vector, which we then use to effectively iterate backward.

Instead, we can skip generating the vector and instead pull directly from the register set the last register. This makes it easier to — for example — iterate through them in pairs.

Reviewed By: alexmalyshev

Differential Revision: D87096974

fbshipit-source-id: 9b2cf1c02e3000fdc7ac086de8c63c43def2add1