Rename STACK_REGISTERS -> DISALLOWED_REGISTERS
Summary: On x86-64 these are the same concepts (the set of disallowed registers is the same set of registers as those that are used for the stack). However on aarch64, we want XZR to also be on the list of disallowed registers, which is not explicitly a stack register, but rather something that has special meaning.

Reviewed By: alexmalyshev

Differential Revision: D87260451

fbshipit-source-id: 78602c1da4f4719d06e5d5d0f14258092bafb90c