Run tests as their own Sandcastle jobs
Summary:
This may or may not be a good idea... please look at the `cinderx_ci-...` runs on this diff and see how you feel about this. With this diff I saw my sandcastle jobs as being throttled - I'm not sure if Sandcastle will adapt to the fact that these run quickly over time. The "Duration" for these on the diff page also seems to include queueing time, so you can see they actually run much faster.

This splits our test cases out to run one by one. Now if a test is failing we'll have just a single failure instead of having an entire opaque test run failing.

We can't read the list of tests at build time so we now have a cached list of known tests. There's a new test which tests that the tests are up to date, and if they're not `TestScripts/test_test_defs.sh` can be run and the resulting `/tmp/tests.txt` can be copied over `cinderx/tests.bzl`.

The test runners needed to be updated to have a new `--list` option to print out all of the tests they discovered. There also were some small updates so that the individual tests would successfully run and pass:
* We need to set sys.path on 3.10 in the worker like we do on 3.12+
* When running an individual module on 3.10 we need to fallback to `_runtest_inner2` so that it can run 'test_main` to setup the module. This also means we need to get rid of it's propensity to pre-pend `test.` if a test name doesn't start with that so that we can run `test_cinderx` tests.
* If no tests are run we report success (because our filters can now filter down to 0 tests).

Probably the most annoying thing about this is when we're onboarding a new version we don't get a nice concise list of test failures to update a failures script. I'm curious what people think about that... we could always have some easy-to-enable thing which runs all of the tests.

This also enables base revision retries now that they should be a little faster as they target individual tests.

Reviewed By: alexmalyshev

Differential Revision: D85780300

fbshipit-source-id: 94168f134ba98d62088eaa8c80f16f3f2cc7231a