Make lightweight frames configurable at runtime
Summary:
Adding a new FrameMode for lightweight frames so we can switch between them and
normal frames on 3.12.

Because there's fewer ENABLE_LIGHTWEIGHT_FRAMES ifdefs more of the code starts
compiling for 3.14.  This diff fixes some of it but doesn't address all of it.
There's still uses of ENABLE_LIGHTWEIGHT_FRAMES so it's not 100% a runtime option
just yet.

Reviewed By: czardoz

Differential Revision: D83871317

fbshipit-source-id: 93ec1314f697b47423788082d75092a181b619b3