Add cinderx.jit.set_max_code_size
Summary: Add a new API method called `set_max_code_size()` that allows Python developers to set the maximum code size that the JIT can write. The argument passed to the method must be a non-negative int. if the value is 0, this effectively makes the max_code_size threshold infinite (as is already the case).

Reviewed By: alexmalyshev

Differential Revision: D88091014

fbshipit-source-id: 33aa94acedcac719eb5b215235b25fa3d1ee6358