Workaround mimalloc __THROW decl in cpp files
Summary: This fixes an issue with the 3.15 build. When we include the mimalloc header in our cinderx/python.h we end up getting lots of build errors in C++ files. These are from the presence of `__THROW` being defined which mimalloc is checking for and using for `mimalloc_throw`. So we just undefine it away as we don't really care about this.

Reviewed By: jbower-fb

Differential Revision: D88178338

fbshipit-source-id: 2612373dc341599500986d491c58f39e1e6a5715