Fix removeUnreachableInstruction paths that can access dangling pointers
Summary:
I'm kind of shocked that we haven't hit this before, but there's some combination of Python code changes in 3.14 and inlining that cause us to finally start hitting this when we enable lightweight frames in the stack of D87386840.

In `removeUnreachableInstructions` we build up a map of register -> instruction. We then run through and do some processing on those including deleting the instructions that are stored in the map. But we never remove the deleted instructions *from* the map.This leaves a dangling pointer to the freed instructions in the map and when we do the processing of the `instrs_using_reg` map (https://www.internalfb.com/code/fbsource/[dc158b99a97d]/fbcode/cinderx/Jit/hir/pass.cpp?lines=772) we can end up accessing the freed memory and we blow up loudly on ASAN builds.

This adds a new lambda which we use to cleanup the map before we delete any instructions.

Reviewed By: alexmalyshev

Differential Revision: D88116938

fbshipit-source-id: 60b00b65fe5e93c4a4ef4eafb37fd086dc3f8a5c