Check maximum code size before compilation
Summary:
There is a config variable called `max_code_size` that specifies how much memory the JIT can write. We currently [check](https://www.internalfb.com/code/fbsource/[6280df1c86fa49a8c93243f5d8266a5e41126627]/fbcode/cinderx/Jit/pyjit.cpp?lines=2384%2C2394-2397) it during registration. There is a delay, however, between registration and compilation. For example, the first time a function is created it is registered. However, the JIT may only compile the function after N times of the function being called. This means that although we may be below the threshold when we register the function, we may be above the threshold when we compile the function. To fix this potential issue, we check the threshold as well before compilation. Note that there is the chance that we compile the function and end up over the threshold but we'll live with that for now.

This implementation checks the `max_code_size` during `compilePreloader()`. The reason we check it here is because this function is hit in all code paths.
The normal flow that happens during execution is
```
jitVectorcall() -> forcedJitVectorcall() -> tryCompile() -> compileFunction() -> compile_func() -> compilePreloader()
```
However, there is also the flow where `cinderjit.precompile_all()` is called from Python code:
```
precompile_all() -> compile_all()-> multithread_compile_units_preloaded() -> compile_units_preloaded() -> tryCompilePreloaded() -> compilePreloader()
or
precompile_all() -> compile_all()-> compile_units_preloaded() -> tryCompilePreloaded() -> compilePreloader()
```

Reviewed By: alexmalyshev

Differential Revision: D87942083

fbshipit-source-id: 68a1181465bc348fe49e115ff762ad484c81365a